+++ src/azure-cli-core/HISTORY.rst
@@ -4,7 +4,7 @@ Release History
 ===============
 2.0.27
 ++++++
-auth: key on both subscription id and name on msi login
+* auth: key on both subscription id and name on msi login
 * Minor fixes
 
 2.0.26